requirements.txt
tweepy
google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
discord-webhook
python-dotenv
requests
Pillow
discord.py
